[rails model] Adding a User controller and method headers
author: Amy Zhang

Generated a User controller and put in placeholders for the methods that
will be there.
Modified routes.rb file to contain corresponding routes for the
controller methods.
Updated Gemfile to have rspec and pg gems, specified version numbers.
Rails default 'test' folder currently still exists, might delete later
in favor of rspec.
Did not generate a model -- we can do that later once we figure out
exactly what we want to store for each user in the database instead of
having to update the database every time we come up with a new category.