Create basic UI Layout for Shopping List
author: Amy Zhang

This ShoppingListActivity layout consists of a field to add an
ingredient, an add button, and a spinner to select ingredient types from
a predetermined list (which can be easily edited).

In addition to these changes, the shopping_list.xml menu page now has
additional choices in the action bar drop down menu. Currently, these
choices don't lead anywhere but ideally they will be used for
navigation. Furthermore, we would not have login/logout/sign up all on
the same page - it will be dependent on the state of the user, but for
exploration purposes, they all exist together.

Lastly, a variety of new strings now live in strings.xml. These are
mostly the ones designed for the action bar (they all have the form
"action_[name of activity]". There is also a string array for the
ingredient category spinner that can be reused for the addInventory
page.

Hopefully this doesn't break anything!