Defining interface for DatabaseModel and DatabaseModelHelper classes
author: Amy Zhang

Created a DatabaseModelHelper class to extend SQLiteHelper class,
allowing for creation and modification of our local database.
Put in headers for the methods in DatabaseModel that will be used to
interact with the database. These methods will make use of the
DatabaseModelHelper class.
We now have to figure out how to create a database (using SQL code) and
once that's completed, we can fill in these methods!