Added search model and storage
The search model is now named SearchTask.java to be consistent since it extends AsyncTask. I included some documentation about how to use it. It's pretty simple, just call new SearchTask().execute(String request_type, String query), where request_type is either recipe for getting a certain recipe when you already have the Yummly ID or search if you want to search for recipes. query is either the recipe ID or the search query. I still need to add some stuff especially for interactions with the UI but it gives a basic idea of what's going on.

SearchTask returns a storage object now, which will probably be convenient to pass in an Intent. The Storage class is just something I made to organize all the Yummly data. There's a Recipe class, SearchResult class, etc. that should make displaying the info easier.